💥 separate comparison scripts in 3 parts
